pyproject.toml
[project]
name = "samcam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.7.0",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.0.1",
    "pywizlight>=0.6.0",
]
